+++ server/src/main/kotlin/io/lionweb/server/ui/LoggingDeltaChannel.kt
@@ -1,4 +1,4 @@
-package io.lionweb.server
+package io.lionweb.server.ui
 
 import com.google.gson.Gson
 import io.lionweb.client.delta.DeltaChannel
