+++ server/src/main/kotlin/io/lionweb/server/ui/MessageLog.kt
@@ -1,4 +1,4 @@
-package io.lionweb.server
+package io.lionweb.server.ui
 
 import java.util.concurrent.CopyOnWriteArrayList
